src/api/requirements.txt
dnspython==2.3.0
exceptiongroup==1.2.0
fastapi==0.104.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.0
referencing==0.31.0
slugify==0.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
redis==5.0.1
celery==5.3.6
requests==2.31.0
python-dotenv==1.0.0
motor==3.3.2
httpx==0.25.2
sentry-sdk==1.38.0
slowapi==0.1.8
aioboto3==12.1.0
boto3==1.33.1
tiktoken==0.5.2
email-validator==2.1.0.post1
python-multipart==0.0.6
pytest==7.4.3
selenium==4.16.0
websocket_client==1.7.0
python-jose==3.3.0
langchain==0.1.0
cryptography==41.0.7
ruff==0.0.122
svix==1.16.0
chevron==0.14.0
troposphere==4.6.0
beautifulsoup4==4.12.3
simplejson==3.19.2
magika==0.5.1